Added structured logging of request and response
We assign a UUID to each request using middleware. Then,
Using a custom version of fastapi's APIRoute, we log the
UUID, operation ID, JWT claimset, and details of the
request and response as JSON using a custom Python
logger. Query and path parameters are captured, as is the
JSON body of a request (when method is POST, PUT, PATCH).
The JSON body of the response is captured only when the
method is POST, PUT, or PATCH.

Additional work needs to be done in the deployment config
for this app to permanently capture the audit log as it
will be ephemeral with the container process in the
current App Runner deployment.