Added support for pulling a refresh token
Pulling a refresh token requires that the Tapis client be configured with a client ID and Key. These
are provided in config.Settings.tapis_client_id and config.Settings.tapis_client_key. The auth/create_token
function now configures the Tapis client with a) values provided by the server or b) values provided by the
user in the auth/token form.